data type of lyrics duration changed from int to double

+++ app/src/main/java/io/github/abhishekabhi789/lyricsforpoweramp/model/Lyrics.kt
@@ -11,12 +11,12 @@ data class Lyrics(
     @SerializedName("albumName") val albumName: String,
     @SerializedName("plainLyrics") val plainLyrics: String?,
     @SerializedName("syncedLyrics") val syncedLyrics: String?,
-    @SerializedName("duration") val duration: Int
+    @SerializedName("duration") val duration: Double
 ) {
     fun getFormattedDuration(): String {
-        val hours = duration / 3600
-        val minutes = (duration % 3600) / 60
-        val seconds = duration % 60
+        val hours = (duration / 3600).toInt()
+        val minutes = ((duration % 3600) / 60).toInt()
+        val seconds = (duration % 60).toInt()
         val formatted = buildString {
             if (hours > 0) append("$hours:")
             append(String.format(Locale.getDefault(), "%02d:%02d", minutes, seconds))
